# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Sprinkie API
  description: >
    Contract source of truth for the Sprinkie API. Go server stubs
    (oapi-codegen) and the app's TypeScript client types (openapi-typescript)
    are both generated from this file; regenerate whenever it changes.
  version: 0.1.0
paths:
...
  /me:
    post:
      operationId: provisionMe
      summary: Provision and return the signed-in User's world
      description: >
        Idempotent. On first sign-in it atomically creates the User, their
        Journal, and the five seeded categories; on every later call it
        returns the same world unchanged. A deactivated account answers 403
        here and everywhere else (#15); restoring it takes the deliberate
        POST /me/reactivate, never a mere session restore.
      security:
        - bearerAuth: []
      requestBody:
        required: false
        description: >
          Optional one-time seeding hint (#36): the App Language in effect
          at signup names the five Starter Categories. Absent or zh-TW
          seeds the Traditional Chinese set; en seeds Work, Exercise, Food,
          Travel, Personal with the same icons and colors. Only the first
          provisioning seeds — an account that already has Categories is
          never reseeded or renamed, whatever language later calls carry.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProvisionMe"
      responses:
        "200":
          description: The signed-in User's provisioned world.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Me"
        "401":
          description: Missing or invalid access token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: The account is deactivated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Provisioning failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: deactivateMe
      summary: Deactivate the signed-in User's account
      description: >
        Leaving is a right (#15). Deactivation is immediate — every route
        except POST /me answers 403 from here on — and the permanent purge
        follows after a 30-day grace, cascading through Entries, Categories,
        and stored photo files. Signing in again before the purge
        reactivates the account intact. Idempotent.
      security:
        - bearerAuth: []
      responses:
        "204":
          description: The account is deactivated.
        "401":
          description: Missing or invalid access token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Deactivation failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /me/reactivate:
    post:
      operationId: reactivateMe
      summary: Restore a deactivated account within its grace period
      description: >
        The deliberate half of reactivation (#15): the user confirms the
        restore and gets their world back intact. A separate operation from
        POST /me on purpose — the app provisions on every launch, and an
        auto-reactivating /me would let any still-signed-in device cancel a
        deletion silently. Idempotent on active accounts.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: The restored (or already active) User's world.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Me"
        "401":
          description: Missing or invalid access token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Reactivation failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase Auth access token, verified against the project JWKS.
  schemas:
    ProvisionMe:
      type: object
      properties:
        language:
          type: string
          enum: ["zh-TW", "en"]
          description: The client's resolved App Language at signup.
    Me:
      type: object
      required:
        - userId
        - journalId
        - categories
      properties:
        userId:
          type: string
        journalId:
          type: string
        categories:
          type: array
          items:
            $ref: "#/components/schemas/Category"
    Category:
      type: object
      required:
        - id
        - name
        - color
        - icon
        - position
      properties:
        id:
          type: string
        name:
          type: string
        color:
          type: string
          description: Hex color; presets map to the category palette.
        icon:
          type: string
        parentId:
          type: string
        position:
          type: integer
        inUse:
          type: boolean
          description: Whether any Entry references it, as category or refinement.
        hasChildren:
          type: boolean
          description: Whether any Subcategory points at it.
...
    Error:
      type: object
      required:
        - message
      properties:
        message:
          type: string